requirements.txt
# Core
typer>=0.9.0
rich>=13.0.0
python-dotenv>=1.0.0
requests>=2.31.0
PyYAML>=6.0.1
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Scraper
feedparser>=6.0.10
beautifulsoup4>=4.12.0
lxml>=4.9.0  # Secure XML parsing

# Database
sqlalchemy>=2.0.0

# AI / Engines
openai>=1.3.0
anthropic>=0.7.0
ollama>=0.1.0
google-genai>=0.1.0
groq>=0.4.0
transformers>=4.35.0

# Notification
slack-sdk>=3.26.0


mistralai>=1.0.0
